LPS-125155 Separate search container component
