+++ modules/apps/change-tracking/change-tracking-web/src/main/resources/META-INF/resources/publications/css/ChangeTrackingIndicator.scss
